news_fact_checker/news_fact_checker/requirements.txt
crewai
crewai-tools
exa-py
newsapi-python
python-dotenv
requests
pydantic